example/producer/requirements.txt
blinker==1.9.0
click==8.2.1
Deprecated==1.2.18
Flask==3.0.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
json_logger==0.0.1
MarkupSafe==3.0.2
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
pika==1.3.2
structlog==25.4.0
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
